requirements.txt
jsonschema>=4.0.0
google-generativeai>=0.8.0
